one/admin/requirements.txt
asgiref==3.3.4
autopep8==1.5.6
backcall==0.2.0
boto3==1.17.79
botocore==1.20.79
decorator==5.0.7
Django==3.2
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-extensions==3.1.2
django-js-asset==1.2.2
django-storages==1.11.1
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
jmespath==0.10.0
jsonfield==3.1.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycodestyle==2.7.0
Pygments==2.8.1
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
